requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
langchain-google-genai>=0.0.4
openai>=1.3.0
google-generativeai>=0.3.0
replicate>=0.22.0
pyyaml>=6.0.1
python-dotenv>=1.0.0 